refactor(StringUtils): Move StringUtils to domain
