+++ src/test/java/racingcar/utils/StringUtilsTest.java
@@ -3,6 +3,7 @@ package racingcar.utils;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.Test;
+import racingcar.domain.StringUtils;
 
 class StringUtilsTest {
